Added safe saving and retrieval of Exposures to local db

+++ app/src/main/java/com/example/cmp354_covidtracer/Location.java
@@ -1,12 +1,12 @@
 package com.example.cmp354_covidtracer;
 
 public class Location {
-    private int lng;
-    private int lat;
+    private double lng;
+    private double lat;
     //Timestamp
     private int ts;
 
-    public Location(int lng, int lat, int ts) {
+    public Location(double lng, double lat, int ts) {
         this.lng = lng;
         this.lat = lat;
         this.ts = ts;
@@ -15,19 +15,19 @@ public class Location {
     public Location() {
     }
 
-    public int getLng() {
+    public double getLng() {
         return lng;
     }
 
-    public void setLng(int lng) {
+    public void setLng(double lng) {
         this.lng = lng;
     }
 
-    public int getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(int lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
